+++ src/current-line-scroller.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { ViewPort } from './metago.lib';
+import { ViewPort } from './metago.core';
 
 // todo: test revealLine command: 
 // https://code.visualstudio.com/api/references/commands